+++ internal/bubbles/shared/navigator/statusbar/events.go
