+++ internal/bubbles/shared/navigator/statusbar/model.go
